requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0
PyJWT>=2.8.0
bcrypt>=4.2.0
pytest>=8.0
httpx>=0.27
